Add barebones CMake support
This makes EB buildable with CMake on:

- Linux/*BSD
- Windows (EBNET unsupported for now)
- macOS

CMake-based projects can add_subdirectory()/FetchContent*() this and
link eb::eb. If zlib is not found, it is implicitly fetched and built.

The goal is painless embedding into applications. Installation is not
implemented ATM. If you need to "make install" things, use the autoconf
build instead.